# api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Flux API
  version: "1.0.0"
  description: >
    Flux lets you meter continuous, stoppable-anytime payments over the
    Lightning Network — a taxi meter for money. Start a session against a
    receiver's Lightning node, and Flux fires one keysend payment per tick
    interval for as long as the session runs. Stop it at any moment and
    billing halts within one tick, guaranteed: no payment is ever sent
    after a stop is acknowledged.


    Flux ships with a **simulated Lightning backend** by default, so you
    can build a complete integration — start a session, watch it bill,
    stop it, replay the payment history — without running a real
    Lightning node. Point it at a real LND node
    (`LIGHTNING_PROVIDER=lnd`) when you're ready to go live.
  contact:
    name: Flux Developer Support
    url: https://flux.dev/support
  license:
    name: MIT
servers:
  - url: https://api.flux.dev
    description: Hosted API
  - url: http://localhost:8081
    description: Local development
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/webhooks:
    post:
      tags: [Webhooks]
      summary: Register a webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url: { type: string, format: uri, example: "https://yourapp.com/webhooks/flux" }
                events:
                  type: array
                  items:
                    type: string
                    enum:
                      [session.started, payment.sent, payment.failed, session.stopped, session.autostopped]
      responses:
        "201":
          description: >
            Webhook registered. The `secret` field is shown exactly once —
            store it to verify the `Flux-Signature` header on incoming
            deliveries.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id: { type: string }
                      url: { type: string }
                      events: { type: array, items: { type: string } }
                      secret: { type: string, example: "whsec_..." }
                      created_at: { type: string, format: date-time }
    get:
      tags: [Webhooks]
      summary: List webhooks
      responses:
        "200":
          description: Registered webhooks (secrets not included).
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        url: { type: string }
                        events: { type: array, items: { type: string } }
                        disabled: { type: boolean }
                        created_at: { type: string, format: date-time }
  /v1/webhooks/{id}:
    delete:
      tags: [Webhooks]
      summary: Disable a webhook
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Webhook disabled.
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: >
        Pass your API key as a bearer token:
        `Authorization: Bearer flx_test_...`. An `X-API-Key` header is
        also accepted.
...